+++ packages/editor-core/src/ui/ChromeExpanded/index.tsx
@@ -40,7 +40,6 @@ import ToolbarInlineCode from '../ToolbarInlineCode';
 import ToolbarImage from '../ToolbarImage';
 import ToolbarMedia from '../ToolbarMedia';
 import ToolbarTextColor from '../ToolbarTextColor';
-import TableFloatingControls from '../TableFloatingControls';
 import TableFloatingToolbar from '../TableFloatingToolbar';
 import {
   Container,
@@ -324,14 +323,6 @@ export default class ChromeExpanded extends PureComponent<Props, State> {
               popupsMountPoint={popupsMountPoint}
               popupsBoundariesElement={popupsBoundariesElement}
             /> : null}
-
-          {pluginStateTable && !disabled &&
-            <TableFloatingControls
-              pluginState={pluginStateTable}
-              editorView={editorView}
-              popupsMountPoint={this.maxHeightContainer}
-              popupsBoundariesElement={this.maxHeightContainer}
-            /> }
 
           {pluginStateTable && !disabled &&
             <TableFloatingToolbar